requirements.txt
alembic==1.0.1
aniso8601==4.0.1
apt-xapian-index==0.47
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.3.1
cupshelpers==1.0
distro-info==0.18
entrypoints==0.2.3.post3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Guake==3.3.0
httplib2==0.11.3
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
keyring==15.1.0
keyrings.alt==3.1
language-selector==0.1
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
olefile==0.46
pbr==4.2.0
pexpect==4.6.0
Pillow==5.2.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.30.1
PyJWT==1.6.4
python-apt==1.7.0
python-dateutil==2.7.5
python-debian==0.1.33
python-editor==1.0.3
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.6
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.2.12
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
Werkzeug==0.14.1
xkit==0.0.0
